+++ report-engine/src/desktopMain/kotlin/org/octavius/report/HelperClasses.kt
@@ -14,7 +14,7 @@ data class ReportPaginationState(
     val currentPage: Long = 0,
     val totalPages: Long = 1,
     val totalItems: Long = 0,
-    val pageSize: Int = 10
+    val pageSize: Long = 10
 ) {
     fun resetPage(): ReportPaginationState = this.copy(currentPage = 0)
 }
